test: support nan results in dataset.
Also split the test_bem_problems_and_results into test_io_xarray.
